backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
groq==0.11.0
PyPDF2==3.0.1
sentence-transformers==3.1.1
faiss-cpu>=1.12.0
transformers>=4.40.0
torch>=2.0.0
sse-starlette==2.1.3
pydantic==2.9.2
